clay_doll_agent/pyproject.toml
[project]
name = "clay-doll-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.19.0",
]
